+++ android/app/src/main/kotlin/com/cryptork/arena/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.arena.arena
+package com.cryptork.arena
 
 import io.flutter.embedding.android.FlutterActivity